Delete button to remove NFA from Project
