+++ nfafrontend/src/app/current-project/assign-nfa/assign-nfa.component.html
@@ -1,10 +1,10 @@
 <div class="row">
-<div class="col-md-5">
+<div class="col-md-6">
   <app-nfa-list></app-nfa-list>
 
 </div>
 
-<div class="col-md-7">
+<div class="col-md-6">
   <router-outlet></router-outlet>
 </div>
