+++ nfafrontend/src/app/current-project/assign-nfa/nfa-list/nfa-list.component.html
@@ -1,6 +1,6 @@
 <div class="row">
   <div class="col-xs-12" >
-    <button type="button" class="btn btn-danger" (click)="onCancel()">Project: {{project.customerName}}</button>
+    <button type="button" class="btn btn-default" (click)="onCancel()">Project: {{project.customerName}}</button>
     <button class="btn btn-success pull-right" (click)="AssignNfa()" > Assign NFA </button>
   </div>
 </div>
@@ -12,11 +12,14 @@
 </div>
 <hr>
 <div class="row">
-  <div class="col-xs-12">
-
+  <div  *ngFor="let nfaCatalogEl of nfaCatalogs let i = index">
+  <div class="col-xs-10">
     <app-nfa-item
-      *ngFor="let nfaCatalogEl of nfaCatalogs let i = index"
       [nfaCatalog]="nfaCatalogEl"
       [index]="i"></app-nfa-item>
+  </div>
+    <div class="col-xs-2">
+      <button type="button" class="btn btn-danger" style="margin-top: 10px" (click)="onAddOrRemove(nfaCatalogEl)">X</button>
+    </div>
   </div>
 </div>
