+++ nfafrontend/src/app/current-project/assign-nfa/nfa-list/nfa-list.component.ts
@@ -5,6 +5,7 @@ import {DataStorageService} from '../../../shared/data-storage.service';
 import {NfaCatalogModel} from '../../../shared/nfaCatalog.model';
 import {Project} from '../../../shared/project.model';
 import {Subscription} from 'rxjs/Subscription';
+import {Response} from '@angular/http';
 
 @Component({
   selector: 'app-nfa-list',
@@ -48,6 +49,23 @@ export class NfaListComponent implements OnInit {
 
   onCancel(){
       this.router.navigate(['../'], {relativeTo: this.route});
+  }
+
+  onAddOrRemove(nfa: NfaCatalogModel){
+    const editedProject = this.currentProjectService.getProject(this.id);
+    const ind = editedProject.projectNfas.findIndex((x) => x.nfaCatalogId === nfa.nfaCatalogId)
+    if (ind !== -1) {
+      editedProject.projectNfas.splice(ind,1);
+    } else {
+      editedProject.projectNfas.push(nfa);
+    }
+    this.currentProjectService.updateProject(this.id, editedProject);
+    this.dataStorageService.updateProject(editedProject)
+      .subscribe(
+        (response: Response) => {
+          this.currentProjectService.projectsChanged.next(this.currentProjectService.getProjects());
+        }
+      );
   }
 
 }
