+++ nfafrontend/src/app/current-project/project-detail/project-detail.component.html
@@ -1,6 +1,6 @@
 <div class="row">
   <div class="col-xs-12" style="text-align: center">
-    <span class="label label-default">Project Details</span>
+    <small><span class="label label-default">Project Details</span></small>
   </div>
 </div>
 <div class="row">
