+++ nfafrontend/src/app/nfacatalog/nfacatalog-criteria/nfacatalog-criteria-list/nfacatalog-criteria-list.component.html
@@ -1,6 +1,6 @@
 <div class="row">
   <div class="col-xs-12" style="text-align: center">
-    <span class="label label-default">{{nfaFactor.factor}} - Criterias</span>
+    <small><span class="label label-default">{{nfaFactor.factor}} - Criterias</span></small>
   </div>
 </div>
 <hr>
