+++ nfafrontend/src/app/nfacatalog/nfacatalog-criteria/nfacatalog-metric/nfacatalog-metric.component.html
@@ -4,12 +4,12 @@
 
 <div class="row" *ngIf ="onView">
   <div class="col-xs-12">
-    <button class="btn btn-default btn-lg col-xs-1 col-xs-offset-1" (click)="onPrev()" [disabled] = "metricIdx == 0"> < </button>
-    <button class="btn btn-default btn-lg col-xs-4 col-xs-offset-2"
+    <button class="btn btn-default btn-lg col-xs-1" (click)="onPrev()" [disabled] = "metricIdx == 0"> < </button>
+    <button class="btn btn-default btn-lg col-xs-10"
             style="text-align: center"
             (click)="onView = !onView">
       {{nfaMetrics[metricIdx].bezeichnung}}</button>
-    <button class="btn btn-default btn-lg col-xs-1 col-xs-offset-2" (click)="onNext()" [disabled] = "metricIdx == nfaMetrics.length-1"> > </button>
+    <button class="btn btn-default btn-lg col-xs-1" (click)="onNext()" [disabled] = "metricIdx == nfaMetrics.length-1"> > </button>
   </div>
 </div>
