+++ nfafrontend/src/app/nfacatalog/nfacatalog-list/nfacatalog-list.component.html
@@ -1,6 +1,6 @@
 <div class="row">
   <div class="col-xs-12" style="text-align: center">
-    <span class="label label-default">NFA Factors</span>
+    <small><span class="label label-default">NFA Factors</span></small>
   </div>
 </div>
 <hr>
